move recordsets resources to their own file
