+++ vinyldns/resources.go
@@ -50,83 +50,6 @@ type ListFilter struct {
 	MaxItems   int
 }
 
-// RecordSetChange represents a record
-// set change.
-type RecordSetChange struct {
-	Zone       Zone      `json:"zone"`
-	RecordSet  RecordSet `json:"recordSet"`
-	UserID     string    `json:"userId"`
-	ChangeType string    `json:"changeType"`
-	Status     string    `json:"status"`
-	Created    string    `json:"created"`
-	ID         string    `json:"id"`
-}
-
-// RecordSet represents a DNS record set.
-type RecordSet struct {
-	ID           string   `json:"id,omitempty"`
-	ZoneID       string   `json:"zoneId"`
-	OwnerGroupID string   `json:"ownerGroupId,omitempty"`
-	Name         string   `json:"name,omitempty"`
-	Type         string   `json:"type"`
-	Status       string   `json:"status,omitempty"`
-	Created      string   `json:"created,omitempty"`
-	Updated      string   `json:"updated,omitempty"`
-	TTL          int      `json:"ttl"`
-	Account      string   `json:"account"`
-	Records      []Record `json:"records"`
-}
-
-// RecordSetUpdateResponse represents
-// a JSON response from the record set update endpoint.
-type RecordSetUpdateResponse struct {
-	Zone      Zone      `json:"zone"`
-	RecordSet RecordSet `json:"recordSet"`
-	ChangeID  string    `json:"id"`
-	Status    string    `json:"status"`
-}
-
-// Record represents a DNS record
-type Record struct {
-	Address     string `json:"address,omitempty"`
-	CName       string `json:"cname,omitempty"`
-	Preference  int    `json:"preference,omitempty"`
-	Exchange    string `json:"exchange,omitempty"`
-	NSDName     string `json:"nsdname,omitempty"`
-	PTRDName    string `json:"ptrdname,omitempty"`
-	MName       string `json:"mname,omitempty"`
-	RName       string `json:"rname,omitempty"`
-	Serial      int    `json:"serial,omitempty"`
-	Refresh     int    `json:"refresh,omitempty"`
-	Retry       int    `json:"retry,omitempty"`
-	Expire      int    `json:"expire,omitempty"`
-	Minimum     int    `json:"minimum,omitempty"`
-	Text        string `json:"text,omitempty"`
-	Priority    int    `json:"priority,omitempty"`
-	Weight      int    `json:"weight,omitempty"`
-	Port        int    `json:"port,omitempty"`
-	Target      string `json:"target,omitempty"`
-	Algorithm   string `json:"algorithm,omitempty"`
-	Type        string `json:"type,omitempty"`
-	Fingerprint string `json:"fingerprint,omitempty"`
-}
-
-// RecordSetResponse represents the JSON
-// response from the record set endpoint.
-type RecordSetResponse struct {
-	RecordSet RecordSet `json:"recordSet"`
-}
-
-// RecordSetsResponse represents the JSON
-// response from the record sets endpoint.
-type RecordSetsResponse struct {
-	NextID           string      `json:"nextId,omitempty"`
-	MaxItems         int         `json:"maxItems,omitempty"`
-	StartFrom        string      `json:"startFrom,omitempty"`
-	RecordNameFilter string      `json:"recordNameFilter,omitempty"`
-	RecordSets       []RecordSet `json:"recordSets"`
-}
-
 // User represents a vinyldns user.
 type User struct {
 	ID        string `json:"id,omitempty"`
